Let small-chunk debris keep its speed
The "bunch of little chunks" line in five prop explosions computes its speed
with int arithmetic, so any dmg below 100 truncates to 0 and the chunks drop
straight down while the corner debris from the same explosion scatter
normally.

The two sibling assignments in each of these functions are written
correctly, three lines above, with the same variable and the cast this one
omits -- so the intent is not in question.

Stock maps are unaffected: every one of these props spawns with dmg 0, where
all three lines already agree. Only a mapper-set dmg diverges.

+++ src/g_joe_misc.c
@@ -48,7 +48,7 @@ void metal_explode (edict_t *self)
 	ThrowDebris (self, "models/props/metal/metal5.md2", spd, org);
 
 	// a bunch of little chunks
-	spd = 2 * self->dmg / 200;
+	spd = 2.0 * (float)self->dmg / 200.0;
 	org[0] = self->s.origin[0] + crandom() * self->size[0];
 	org[1] = self->s.origin[1] + crandom() * self->size[1];
 	org[2] = self->s.origin[2] + crandom() * self->size[2];
@@ -701,7 +701,7 @@ void radio_explode (edict_t *self)
 	ThrowDebris (self, "models/props/metal/metal5.md2", spd, org);
 
 	// a bunch of little chunks
-	spd = 2 * self->dmg / 200;
+	spd = 2.0 * (float)self->dmg / 200.0;
 	org[0] = self->s.origin[0] + crandom() * self->size[0];
 	org[1] = self->s.origin[1] + crandom() * self->size[1];
 	org[2] = self->s.origin[2] + crandom() * self->size[2];
@@ -1057,7 +1057,7 @@ void trashcanA_explode (edict_t *self)
 	ThrowDebris (self, "models/props/metal/metal5.md2", spd, org);
 
 	// a bunch of little chunks
-	spd = 2 * self->dmg / 200;
+	spd = 2.0 * (float)self->dmg / 200.0;
 	org[0] = self->s.origin[0] + crandom() * self->size[0];
 	org[1] = self->s.origin[1] + crandom() * self->size[1];
 	org[2] = self->s.origin[2] + crandom() * self->size[2];
@@ -1148,7 +1148,7 @@ void crate_stuff (edict_t *self)
 	ThrowDebris_stuff (self, "models/props/stuff/piece2.md2", spd, org);
 
 	// a bunch of little chunks
-	spd = 10 * self->dmg / 200;
+	spd = 10.0 * (float)self->dmg / 200.0;
 	org[0] = self->s.origin[0] + crandom() * self->size[0];
 	org[1] = self->s.origin[1] + crandom() * self->size[1];
 	org[2] = self->s.origin[2] + crandom() * self->size[2];
@@ -4646,7 +4646,7 @@ void vent_explode (edict_t *self)
 	ThrowDebris (self, "models/props/metal/metal5.md2", spd, org);
 
 	// a bunch of little chunks
-	spd = 2 * self->dmg / 200;
+	spd = 2.0 * (float)self->dmg / 200.0;
 	org[0] = self->s.origin[0] + crandom() * self->size[0];
 	org[1] = self->s.origin[1] + crandom() * self->size[1];
 	org[2] = self->s.origin[2] + crandom() * self->size[2];
